backend/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
chromadb
python-multipart
pypdf
python-docx
pydantic
google-generativeai
